refactor(runloop) change the order of phase and ctx argument on plugins_iterator
Phase can never be nil so it makes sense to pass it as a second argument
and not as a third argument.

When merged with the commit that fixes the issue of init_worker not being
called on plugins that implement it, the third argument will be nil
on init_worker.